Update ModEntry.cs
The previous code would replace the addedSpeed of the player instead adding on it, so it'll ignore the addedSpeed from Buffs

I believe these proposed changes should make it work more like a Buff does, adding and removing addedSpeed when applied or removed, without having to use a buff like some mods do.  

It should also become compatible with any speed buffs from other mods like the speed cheat from CJB Cheats Menu, and with any mod that changes the base speed when mounted on a horse, as long as it isn't the same way as a buff adds speed.